Update 13_interface.kt
interface allows multiple inheritance. it is the responsibility of the subclass to implement the methods in the interfaces. if both interface have the same methods, the subclass can should override those to create a new method with the same name, or use super to use the method of the specific interface that was chosen.